+++ src/kernel/lib/lib.h
@@ -3,7 +3,7 @@
 
 #include <stddef.h>
 #include <stdint.h>
-#include <kernel/multiboot.h>
+#include <multiboot.h>
 
 /**
  * Copy n data from src to dest
@@ -12,7 +12,7 @@
  * @param count The number of bytes to be copied (src)
  * @return The modified dest pointer
  */
-void *memcpy(void *dest, const void *src, size_t count);
+void *memcpy(void *dest, const void *src, u32 count);
 
 /**
  * Replace n bytes of dest by val
@@ -21,7 +21,7 @@ void *memcpy(void *dest, const void *src, size_t count);
  * @param count The number of times val should replace dest entry
  * @return The modified dest pointer
  */
-void *memset(void *dest, char val, size_t count);
+void *memset(void *dest, char val, u32 count);
 
 /**
  * Compare the first n bytes of a and b
@@ -30,7 +30,7 @@ void *memset(void *dest, char val, size_t count);
  * @param size The number of bytes to be compared
  * @return -1 if a < b, 0 if a = b and 1 if a > b
  */
-int memcmp(const void *a_ptr, const void *b_ptr, size_t size);
+int memcmp(const void *a_ptr, const void *b_ptr, u32 size);
 
 void memory_info_init(struct multiboot_tag_basic_meminfo *tag);
 
@@ -38,6 +38,6 @@ void memory_mmap_init(struct multiboot_tag_mmap *tag);
 
 void memory_print();
 
-uint32_t memory_get_all();
+u32 memory_get_all();
 
 #endif
