+++ src/kernel/lib/string/strcat.c
@@ -1,11 +1,11 @@
-#include <kernel/lib/string.h>
+#include <lib/string.h>
 
 void strcat(char *dest, const char *orig)
 {
-	size_t s_dest = strlen(dest);
-	size_t s_orig = strlen(orig);
+	u32 s_dest = strlen(dest);
+	u32 s_orig = strlen(orig);
 
-	for (size_t i = 0; i < s_orig; i++)
+	for (u32 i = 0; i < s_orig; i++)
 		dest[s_dest + i] = orig[i];
 	dest[s_dest + s_orig] = 0;
 }
