+++ src/kernel/lib/string/strcmp.c
@@ -1,11 +1,11 @@
-#include <kernel/lib/string.h>
+#include <lib/string.h>
 
 char strcmp(const char *a, const char *b)
 {
 	if (strlen(a) != strlen(b))
 		return 1;
 
-	for (size_t i = 0; i < strlen(a); i++)
+	for (u32 i = 0; i < strlen(a); i++)
 		if (a[i] != b[i])
 			return 1;
