+++ src/kernel/lib/string/strdisp.c
@@ -1,8 +1,8 @@
-#include <kernel/lib/string.h>
+#include <lib/string.h>
 
 void strdisponce(char *str)
 {
-	for (size_t i = sizeof(str) + 2; i > 0; i--)
+	for (u32 i = sizeof(str) + 2; i > 0; i--)
 		str[i] = str[i - 1];
 	str[0] = 0;
 }
