+++ src/kernel/lib/string/strdup.c
@@ -1,9 +1,9 @@
-#include <kernel/lib/string.h>
-#include <kernel/memory/alloc.h>
+#include <lib/string.h>
+#include <memory/alloc.h>
 
 char *strdup(const char *orig)
 {
-	size_t s_orig = strlen(orig);
+	u32 s_orig = strlen(orig);
 	char *ret = (char *)kmalloc(s_orig + 1);
 	strcpy(ret, orig);
 	return ret;
